pyproject.toml
[project]
name = "chatbot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "print-color>=0.4.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
